repo reorg: Modularize InSAR plotting functions
