+++ src/nisarqa/__init__.py
@@ -109,3 +109,4 @@ from .parameters import *
 from .utils import *
 from .products import *
 from .validate import *
+from .processing import *
